Get rid of interface duplication by removing AuthManager and ResponseCallback
