+++ Atlas/android/app/src/main/java/com/bounswe2017/group10/atlas/auth/LoginFragment.java
@@ -13,7 +13,11 @@ import android.widget.Toast;
 import com.bounswe2017.group10.atlas.R;
 import com.bounswe2017.group10.atlas.httpbody.LoginRequest;
 import com.bounswe2017.group10.atlas.httpbody.LoginResponse;
-import com.bounswe2017.group10.atlas.remote.ResponseCallback;
+import com.bounswe2017.group10.atlas.remote.APIUtils;
+
+import retrofit2.Call;
+import retrofit2.Callback;
+import retrofit2.Response;
 
 
 public class LoginFragment extends Fragment {
@@ -65,28 +69,27 @@ public class LoginFragment extends Fragment {
             loginRequest.setPassword(pw);
 
             // send async login request
-            AuthManager.login(loginRequest, new OnLoginSuccess(), new OnLoginFailure());
+            APIUtils.getAPI().login(loginRequest).enqueue(new OnLoginResponse());
         });
         return view;
     }
 
     /**
-     * Implement the tasks to perform upon successful log in request.
+     * Implement retrofit response callback interface to be used for login requests.
      */
-    class OnLoginSuccess implements ResponseCallback<LoginResponse> {
-        public void onResponse(LoginResponse response) {
-            // handle login success view updates
-            Toast.makeText(getActivity().getApplicationContext(), "Successfully logged in", Toast.LENGTH_LONG).show();
+    private class OnLoginResponse implements Callback<LoginResponse> {
+        @Override
+        public void onResponse(Call<LoginResponse> call, Response<LoginResponse> response) {
+            if (response.isSuccessful()) {
+                Toast.makeText(getActivity().getApplicationContext(), "Successfully logged in", Toast.LENGTH_LONG).show();
+            } else {
+                Toast.makeText(getActivity().getApplicationContext(), "Couldn't log in", Toast.LENGTH_LONG).show();
+            }
         }
-    }
 
-    /**
-     * Implement the tasks to perform upon failed log in request.
-     */
-    class OnLoginFailure implements  ResponseCallback<LoginResponse> {
-        public void onResponse(LoginResponse response) {
-            // handle login failure view updates
-            Toast.makeText(getActivity().getApplicationContext(), "Couldn't log in", Toast.LENGTH_LONG).show();
+        @Override
+        public void onFailure(Call<LoginResponse> call, Throwable t) {
+
         }
     }
 }
