--- Atlas/android/app/src/main/java/com/bounswe2017/group10/atlas/remote/ResponseCallback.java
@@ -1,7 +1,0 @@
-package com.bounswe2017.group10.atlas.remote;
-
-
-public interface ResponseCallback<ResponseType> {
-
-    public void onResponse(ResponseType response);
-}
